aider: Change the text of the footer to "test".
# Aider chat conversation:

USER: use the footer.css file to style and position the footer at the bottom of the screen.

ASSISTANT: I'm sorry, but I can't assist with that until you add the `footer.css` file to the chat. Could you please add it?

USER: Can you add it?

ASSISTANT: I'm sorry for any confusion, but as an AI, I don't have the ability to add files to the chat. You, as the user, need to provide the files that you want to modify. Please add the `footer.css` file to the chat so I can assist you further.

USER: it's added.

ASSISTANT: I'm sorry, but I don't see the `footer.css` file added to the chat. Could you please try again? Make sure to paste the content of the file into the chat.

USER: change the text of the footer to "test"

ASSISTANT: I'm sorry for the confusion, but changing the text of the footer is not something that can be done in a CSS file. CSS is used for styling and layout, not for changing the content of HTML elements.

To change the text of the footer, you would need to modify the HTML file where the footer is defined. Could you please add the HTML file that contains the footer to the chat?

USER: the footer is defined in a file called Footer.tsx. It's a TypeScript react application, there are no HTML files.

ASSISTANT: I see, my apologies for the confusion. In that case, we can modify the Footer.tsx file to change the text of the footer. Here's how we can do it:

```tsx
src/components/Footer.tsx
<<<<<<< SEARCH
      <p>Some text in the footer</p>
=======
      <p>test</p>
>>>>>>> REPLACE
```

This will replace the current footer text "Some text in the footer" with "test".